pyproject.toml
[project]
name = "tutor-ingest"
version = "0.1.0"
description = "교재 PDF 인제스트 파이프라인 + 관리 UI + API 서버 + LangGraph 튜터 에이전트"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    "openai>=1.40",
    "psycopg[binary]>=3.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "streamlit>=1.36",
    "pydantic>=2.7",
    "langchain[openai]>=0.3.27",
    "langchain-openai>=0.3.0",
    "langgraph>=0.6.6",
]

[dependency-groups]
dev = ["pytest>=8", "ipykernel>=6"]

[tool.uv]
dev-dependencies = ["pytest>=8", "ipykernel>=6"]
